// src/dto/cartRequest.dto.ts
import { Static, Type } from "@sinclair/typebox"; // define and validate types

// This is Schema, necessary for validating incoming Request inputs. Then based on the schema, it generates some kind of static type.
export const CartRequestSchema = Type.Object({
  productId: Type.Integer(),
  // customerId: Type.Integer(),
  qty: Type.Integer(),
});
// Need 1 Static type that can be used to create cart.
export type CartRequestInput = Static<typeof CartRequestSchema>;

export const CartEditRequestSchema = Type.Object({
  id: Type.Integer(),
  qty: Type.Integer(),
});
export type CartEditRequestInput = Static<typeof CartEditRequestSchema>;

type CartLineItem = {
  id: number;
  productId: number;
  itemName: string;
  price: string;
  qty: number;
  variant: string | null;
  createdAt: Date;
  updatedAt: Date;
  availability?: number;
};

export interface CartWithLineItems {
  id: number;
  customerId: number;
  lineItems: CartLineItem[];
  createdAt: Date;
  updatedAt: Date;
}